refactor: separate sync and async transaction managers
- Create async_transaction.py for AsyncSQLiteTransactionManager
- Remove aiosqlite import from transaction.py (sync only)
- Fix CLI --info to call introspect_and_adapt() after connect
- Update imports in __init__.py and async_backend.py

This allows synchronous SQLite backend to work without aiosqlite dependency.
